+++ example/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:light_html_editor/editor.dart';
-import 'package:light_html_editor/placeholder.dart';
+import 'package:light_html_editor/light_html_editor.dart';
 
 void main() {
   runApp(MyApp());