+++ lib/renderer.dart
@@ -1,10 +1,11 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
+import 'package:light_html_editor/api/parser.dart';
 import 'package:light_html_editor/editor.dart';
 import 'package:light_html_editor/placeholder.dart';
-import 'package:light_html_editor/richtext/renderer_properties.dart';
-import 'package:light_html_editor/richtext/richtext_node.dart';
+import 'package:light_html_editor/data/renderer_properties.dart';
+import 'package:light_html_editor/api/richtext_node.dart';
 
 ///
 /// Lightweight HTML renderer converting basic HTML text into Richtext
@@ -127,15 +128,11 @@ class RichtextRenderer extends StatelessWidget {
         full = true;
       }
 
-      for (RichTextPlaceholder placeholder in placeholders) {
-        String search =
-            "$placeholderMarker${placeholder.symbol}$placeholderMarker";
-
-        nodeText = nodeText.replaceAll(
-          RegExp(search),
-          "${placeholder.value}",
-        );
-      }
+      nodeText = Parser().replaceVariables(
+        nodeText,
+        placeholders: placeholders,
+        placeholderMarker: placeholderMarker,
+      );
 
       tmp.add(
         TextSpan(